feat(core): move mergeDOMRects in to the main dom utils
