+++ .changeset/olive-flowers-dress.md
@@ -1,0 +1,6 @@
+---
+'@remirror/core-utils': patch
+'@remirror/extension-react-tables': patch
+---
+
+Expose the `mergeDOMRects` helper from core utils
