+++ packages/remirror__extension-react-tables/src/utils/dom.ts
@@ -9,19 +9,6 @@ export function stopEvent(e: Pick<MouseEvent, 'preventDefault' | 'stopPropagatio
   e.stopPropagation();
 }
 
-/**
- * Merge two DOMRect objects into a one big DOMRect object that contains both two DOMRect objects.
- */
-export function mergeDOMRects(rect1: DOMRect, rect2: DOMRect): DOMRect {
-  const left = Math.min(rect1.left, rect2.left);
-  const right = Math.max(rect1.right, rect2.right);
-  const top = Math.min(rect1.top, rect2.top);
-  const bottom = Math.max(rect1.bottom, rect2.bottom);
-  const width = right - left;
-  const height = bottom - top;
-  return new DOMRect(left, top, width, height);
-}
-
 interface Coord {
   x: number;
   y: number;
